+++ tests/Fixtures/SettingsCluster.php
@@ -1,0 +1,12 @@
+<?php
+
+declare(strict_types=1);
+
+namespace LaBoiteACode\FilamentLogsExplorer\Tests\Fixtures;
+
+use Filament\Clusters\Cluster;
+
+class SettingsCluster extends Cluster
+{
+    //
+}